centralise eslint config
