+++ apps/emotion/package.json
@@ -13,10 +13,12 @@
     "@emotion/cache": "^11.11.0",
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
+    "@repo/eslint-config": "workspace:^",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
+    "@repo/typescript-config": "workspace:^",
     "@types/react": "^18.2.66",
     "@types/react-dom": "^18.2.22",
     "@typescript-eslint/eslint-plugin": "^7.2.0",
@@ -25,7 +27,6 @@
     "eslint": "^8.57.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.6",
-    "@repo/typescript-config": "workspace:^",
     "typescript": "^5.2.2",
     "vite": "5.3.1",
     "vite-plugin-hash-csp": "workspace:^"
