+++ apps/less/package.json
@@ -10,6 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@repo/eslint-config": "workspace:^",
     "react": "^18.3.1",
     "react-dom": "^18.3.1"
   },
