+++ apps/react/package.json
@@ -10,10 +10,12 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@repo/eslint-config": "workspace:^",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
+    "@repo/typescript-config": "workspace:^",
     "@types/react": "^18.2.66",
     "@types/react-dom": "^18.2.22",
     "@typescript-eslint/eslint-plugin": "^7.2.0",
@@ -23,8 +25,7 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.6",
     "typescript": "^5.2.2",
-    "vite-plugin-hash-csp": "workspace:^",
-    "@repo/typescript-config": "workspace:^",
-    "vite": "5.3.1"
+    "vite": "5.3.1",
+    "vite-plugin-hash-csp": "workspace:^"
   }
 }
