+++ packages/vite-plugin-hash-csp/package.json
@@ -41,7 +41,8 @@
     "rollup": "^4.18.0",
     "typescript": "^5.5.2",
     "vite": "5.3.1",
-    "vitest": "^1.6.0"
+    "vitest": "^1.6.0",
+    "@repo/eslint-config": "workspace:^"
   },
   "dependencies": {
     "cheerio": "^1.0.0-rc.12"
